asclite (works on Release)

+++ src/rfilter1/include/calloc2.c
@@ -13,7 +13,7 @@
   if (db_level > 3) printf("%sdoing %s\n",(char *)pdb,(char *)proc);
   x = calloc(nobj,size);
   if (x == NULL) fatal_error(calling_proc,(Char *)"MEM ALLOC",-1);
-  if (memory_trace) printf("%s CALLOC %lx\n",pdb,(long)x);
+  if (memory_trace) printf("%s CALLOC %lx\n",pdb,x);
   return x;
  }
 /* end calloc2.c */
