+++ src/rfilter1/include/strdup2.c
@@ -15,7 +15,7 @@ if (db_level > 3) printf("%sdoing %s\n",pdb,proc);
  pd = (Char*)malloc((size_t)strlen((char *)ps)+1);
  if (pd == NULL) fatal_error(calling_proc,(Char *)"MEM ALLOC",-1);
   else   pd = (Char *)strcpy((char *)pd,(char *)ps);
- if (memory_trace) printf("%s MALLOC %lx\n",pdb,(long)pd);
+ if (memory_trace) printf("%s MALLOC %lx\n",pdb,pd);
   return pd;
  }
 /* end strdup2.c */
